refactor rest-client/utils directory
