+++ packages/rest-client/src/services/promise/updateOne/updateOne.test.ts
@@ -1,7 +1,7 @@
 import { describe, it, expect } from "vitest"
 import { createStore } from "../../store"
 import { updateOne } from "./updateOne"
-import { flattenResourcesIntoRecords } from "../../utils/records"
+import { flattenResourcesIntoRecords } from "../../utils"
 import {
   fakeDataSource,
   testDataRecords,
