+++ packages/rest-client/src/services/utils/index.ts
@@ -1,4 +1,7 @@
 export * from "./meta"
+export * from "./records"
+export * from "./request"
+export * from "./response"
 export * from "./schema"
 export * from "./selector"
-export * from "./records"
+export * from "./unflatten"
