+++ packages/rest-client/src/services/utils/meta.ts
@@ -1,4 +1,4 @@
-import type { Meta, MetaError } from "../../types"
+import type { Meta, MetaError } from "../types"
 
 export function getMeta(
   error: MetaError | undefined,
