--- packages/rest-client/src/services/utils/meta/index.ts
@@ -1,1 +1,0 @@
-export * from "./meta"
