--- packages/rest-client/src/services/utils/records/index.ts
@@ -1,2 +1,0 @@
-export * from "./records"
-export * from "./unflatten"
