--- packages/rest-client/src/services/utils/schema/index.ts
@@ -1,1 +1,0 @@
-export * from "./schema"
