--- packages/rest-client/src/services/utils/selector/index.ts
@@ -1,1 +1,0 @@
-export * from "./selector"
