+++ packages/rest-client/src/services/utils/unflatten.ts
@@ -7,7 +7,7 @@ import type {
   GetSchemaFromName,
   GetSchemaNames,
   UpdateType,
-} from "../../types"
+} from "../types"
 
 export function unflattenData<
   const TSchemas extends Record<string, PartialSchema>,
